+++ IM_PJ/tao-ma-vach.aspx.cs
@@ -275,6 +275,7 @@ namespace IM_PJ
             string[] list = hdfListProduct.Value.Split(';');
             if (list.Count() > 0)
             {
+                var temps = new List<String>();
                 productPrint += "<div class=\"qcode\">";
                 for (int i = 0; i < list.Length - 1; i++)
                 {
@@ -284,8 +285,8 @@ namespace IM_PJ
                     for (int j = 0; j < quantity; j++)
                     {
                         barcodeValue = list2[1];
-
-                        barcodeImage = "/uploads/barcodes/" + barcodeValue + ".png";
+                        var imageName = String.Format("{0}{1}.png", DateTime.UtcNow.ToString("yyyyMMddHHmmss"), Guid.NewGuid());
+                        barcodeImage = "/uploads/barcodes/" + imageName;
 
                         System.Drawing.Image barCode = PJUtils.MakeBarcodeImage(barcodeValue, 2,true);
 
@@ -295,6 +296,8 @@ namespace IM_PJ
                         productPrint += "<div class=\"img\"><img src=\"data:image/png;base64, " + Convert.ToBase64String(File.ReadAllBytes(Server.MapPath("" + barcodeImage + ""))) + "\"></div>";
                         productPrint += "<div><h1>" + barcodeValue + "</h1></div>";
                         productPrint += "</div>";
+
+                        temps.Add(imageName);
                     }
                 }
                 productPrint += "</div>";
@@ -306,7 +309,13 @@ namespace IM_PJ
                 string[] filePaths = Directory.GetFiles(Server.MapPath("/uploads/barcodes/"));
                 foreach (string filePath in filePaths)
                 {
-                    File.Delete(filePath);
+                    foreach (var item in temps)
+                    {
+                        if (filePath.EndsWith(item))
+                        {
+                            File.Delete(filePath);
+                        }
+                    }
                 }
                 ScriptManager.RegisterClientScriptBlock(this, this.GetType(), "script", "$(function () { printBarcode() });", true);
             }
